20201019 mark

+++ monitor_backend/views_hub.py
@@ -17,8 +17,8 @@ import monitor_backend.views_commons as roda
 def hub_info(req):
     res = {"status": "OK", "data": {}}
 
-    sql_get_hub_addr = "SELECT `{}`,`service_port` FROM `network` "\
-        "WHERE `role` = 'hub-edge'".format(
+    sql_get_hub_addr = "SELECT `{}`,`service_port`,`docker_port` "\
+        "FROM `network` WHERE `role` = 'hub-edge'".format(
             "vlan_addr" if roda.Is_overlay_network else "addr")
     try:
         roda.mydb_cursor.execute(sql_get_hub_addr)
@@ -31,7 +31,8 @@ def hub_info(req):
 
     res["data"] = {
         "vlan_addr" if roda.Is_overlay_network else "addr": hub_addr[0],
-        "port": hub_addr[1]
+        "port": hub_addr[1],
+        "docker_port": hub_addr[2]
     }
 
     return HttpResponse(json.dumps(res), content_type="application/json")
@@ -87,34 +88,29 @@ def hub_delete_image(req):
 
     delete_data = json.loads(req.body.decode("UTF-8"))
 
-    if('name' in delete_data and 'tag' in delete_data):
-
-        sql_get_hub_addr = "SELECT `{}`,`service_port` FROM `network` WHERE "\
-            "`role` = 'hub-edge'".format(
-                "vlan_addr" if roda.Is_overlay_network else "addr")
-        try:
-            roda.mydb_cursor.execute(sql_get_hub_addr)
-            hub_addr = roda.mydb_cursor.fetchone()
-        except Exception as e:
-            roda.logger.error("Get Hub (vlan) addr query [{}] failed: [{}]".format(
-                sql_get_hub_addr, e))
-            res["status"] = "Get Hub (vlan) addr query failed: [{}]".format(e)
-
+    if('name' in delete_data and 'tag' in delete_data and
+       "hub_addr" in delete_data and "hub_port" in delete_data):
         try:
             image_name = delete_data["name"].split('/')[-1]
             image_tag = delete_data["tag"]
+            hub_addr = delete_data["hub_addr"]
+            hub_port = delete_data["hub_port"]
+
             roda.logger.info("query {}".format("http://{}:{}/v2/{}/manifests/{}".format(
-                hub_addr[0], hub_addr[1], image_name, image_tag)))
+                hub_addr, hub_port, image_name, image_tag)))
             hub_resp = requests.get(
                 "http://{}:{}/v2/{}/manifests/{}".format(
-                    hub_addr[0], hub_addr[1], image_name, image_tag),
+                    hub_addr, hub_port, image_name, image_tag),
                 headers={"Accept": "application/vnd.docker.distribution.manifest.v2+json"})
             image_digest = hub_resp.headers["Docker-Content-Digest"]
+
             roda.logger.info("image_digest of {}:{} is ".format(
                 image_name, image_tag)+image_digest)
+
             hub_resp = requests.delete(
-                "http://{}:{}/v2/{}/manifests/{}".format(hub_addr[0], hub_addr[1],
-                                                         image_name, image_digest))
+                "http://{}:{}/v2/{}/manifests/{}".format(
+                    hub_addr, hub_port, image_name, image_digest))
+
             if(hub_resp.status_code != 202):
                 raise ValueError("Status Code of delete is not 202")
 
@@ -130,28 +126,35 @@ def hub_delete_image(req):
 
 @dj_http.require_GET
 def hub_list_docker_images(req):
+
+    docker_port = req.GET.get("docker_port")
+
     res = {"status": "OK"}
 
-    sql_get_hub_addr = "SELECT `{}` FROM `network` WHERE"
-    " `role` = 'hub-edge'".format(
-        "vlan_addr" if roda.Is_overlay_network else "addr")
-    try:
-        roda.mydb_cursor.execute(sql_get_hub_addr)
-        hub_addr = roda.mydb_cursor.fetchone()
-    except Exception as e:
-        roda.logger.error("Get Hub (vlan) addr query [{}] failed: [{}]".format(
-            sql_get_hub_addr, e))
-        res["status"] = "Get Hub (vlan) addr query failed: [{}]".format(e)
-        return HttpResponse(json.dumps(res), content_type="application/json")
+    if(docker_port):
 
-    try:
-        docker_cli = dockerClient(
-            base_url='tcp://{}:{}'.format(hub_addr[0], "2375"))
-        res["data"] = docker_cli.images()
-    except Exception as e:
-        errmsg = "Get Hub docker engine images failed: [{}]".format(e)
-        roda.logger.error(errmsg)
-        res["status"] = errmsg
+        sql_get_hub_addr = "SELECT `{}` FROM `network` WHERE"
+        " `role` = 'hub-edge'".format(
+            "vlan_addr" if roda.Is_overlay_network else "addr")
+        try:
+            roda.mydb_cursor.execute(sql_get_hub_addr)
+            hub_addr = roda.mydb_cursor.fetchone()
+        except Exception as e:
+            roda.logger.error("Get Hub (vlan) addr query [{}] failed: [{}]".format(
+                sql_get_hub_addr, e))
+            res["status"] = "Get Hub (vlan) addr query failed: [{}]".format(e)
+            return HttpResponse(json.dumps(res), content_type="application/json")
+
+        try:
+            docker_cli = dockerClient(
+                base_url='tcp://{}:{}'.format(hub_addr[0], docker_port))
+            res["data"] = docker_cli.images()
+        except Exception as e:
+            errmsg = "Get Hub docker engine images failed: [{}]".format(e)
+            roda.logger.error(errmsg)
+            res["status"] = errmsg
+    else:
+        res["status"] = "miss parameter"
 
     return HttpResponse(json.dumps(res), content_type="application/json")
 
@@ -185,10 +188,12 @@ def hub_run_image(req):
         base_rate = post_data["base_rate"]
 
         sql_query = "SELECT `network`.`{}` as `addr`, "\
-                    "`sink_sources`.`docker_port` as `port`, "\
-                    "`sink_sources`.`ap_hostname` as `ap_hostname`, "\
-                    "`sink_sources`.`filter_exist` as `filter_exist` "\
-                    " FROM `sink_sources` "\
+                    "`network`.`docker_port` as `port`, "\
+                    "`sink_sources`.`ap_lan_addr` as `ap_lan_addr`, "\
+                    "`sink_sources`.`filter_exist` as `filter_exist`, "\
+                    "`sink_sources`.`influx_port` as `influx_port`,"\
+                    "`sink_sources`.`domain_map` as `domain_map` "\
+                    "FROM `sink_sources` "\
                     "JOIN `network` ON `sink_sources`.`network_id`=`network`.`id` "\
                     "WHERE `sink_sources`.`id`={}".format(
                         'vlan_addr' if roda.Is_overlay_network else 'addr', target_id)
@@ -198,8 +203,16 @@ def hub_run_image(req):
             result = roda.mydb_cursor.fetchone()
             docker_addr = result[0]
             docker_port = result[1]
-            ap_hostname = result[2]
+            ap_lan_addr = result[2]
             filter_exist = result[3]
+            influx_port = result[4]
+            domain_map = result[5]
+
+            domain_lists = {}
+
+            for d_ip in domain_map.split(','):
+                d_ip_split = d_ip.split(':')
+                domain_lists[d_ip_split[0]] = d_ip_split[1]
 
             if(filter_exist):
                 res["status"] = "Filter exist in the source system"
@@ -216,10 +229,13 @@ def hub_run_image(req):
                     "SINK": "{}".format(sink_info),
                     "SOURCE": "{}".format(source_info),
                     "MAXATTR": max_attr,
-                    "INFLUX": "http://{}:8086".format(ap_hostname)
+                    "INFLUX": "http://{}:8086".format(ap_lan_addr)
                 },
-                extra_hosts={ap_hostname: docker_addr}
+                extra_hosts=domain_lists
             )
+
+            roda.logger.info(
+                "run image success [{}:{}]".format(image_name, image_tag))
 
         except Exception as e:
             roda.logger.error("run image failed [{}]:{}".format(sql_query, e))
@@ -228,17 +244,17 @@ def hub_run_image(req):
 
         try:
             influx_client = InfluxDBClient(
-                docker_addr, "8086", "lab126", "lab126", "roda")
+                docker_addr, influx_port, "lab126", "lab126", "roda")
             influx_client.drop_measurement("dynamicParams")
             influx_client.write_points([{
                 "measurement": "dynamicParams",
                 "tags": {
                 },
                 "fields": {
-                    "maxThreads": max_thread,
+                    "maxThreads": int(max_thread),
                     "baseRate": base_rate,
-                    "matchWinSize": win_size,
-                    "expMatchTime": match_threshold
+                    "matchWinSize": int(win_size),
+                    "expMatchTime": int(match_threshold)
                 }
             }])
         except Exception as e:
@@ -311,7 +327,8 @@ def hub_list_file(req):
     res["data"] = []
     for f in file_list:
         res["data"].append({"name": f,
-                            "url": "http://192.168.0.107/static/"+f})
+                            "url": "http://{}/static/{}".format(
+                                roda.DomainName, f)})
 
     return HttpResponse(json.dumps(res), content_type="application/json")
 
@@ -342,18 +359,22 @@ def hub_upload_dockerfile(req):
     file_obj = req.FILES.get('file', None)
     hub_addr = req.POST.get('addr', None)
     hub_port = req.POST.get('port', None)
+    docker_port = req.POST.get('docker_port', None)
     new_image_name = req.POST.get('new_image_name', None)
     new_image_tag = req.POST.get('new_image_tag', "latest")
 
-    if(file_obj and hub_addr and hub_port and new_image_name):
+    if(file_obj and hub_addr and hub_port and
+       new_image_name and docker_port):
         try:
-            docker_cli = APIClient(base_url="tcp://{}:2375".format(hub_addr))
+            docker_cli = APIClient(
+                base_url="tcp://{}:{}".format(hub_addr, docker_port))
             docker_resp = docker_cli.build(
                 fileobj=BytesIO(file_obj.read()), rm=True, tag="{}:{}/{}:{}".format(
                     hub_addr, hub_port, new_image_name, new_image_tag))
             for line in docker_resp:
                 line_stream_info = line.decode("UTF-8")
-                if("stream" not in line_stream_info):
+                if("stream" not in line_stream_info and
+                   "aux" not in line_stream_info):
                     errmsg = "Build new image '{}:{}' failed: {}".format(
                         new_image_name, new_image_tag,
                         line_stream_info)
@@ -362,7 +383,11 @@ def hub_upload_dockerfile(req):
                     resp = HttpResponse(json.dumps(
                         res), content_type="application/json")
                     resp.status_code = 415
+                    return resp
                     break
+                else:
+                    roda.logger.info(line_stream_info)
+
             roda.logger.info("Build new image '{}:{}' success".format(
                 new_image_name, new_image_tag))
 
@@ -375,11 +400,16 @@ def hub_upload_dockerfile(req):
             resp = HttpResponse(json.dumps(
                 res), content_type="application/json")
             resp.status_code = 415
+            return resp
 
         try:
             docker_resp = docker_cli.push(
                 "{}:{}/{}".format(hub_addr, hub_port, new_image_name),
                 tag=new_image_tag, stream=True, decode=True)
+            for line in docker_resp:
+                roda.logger.info(line)
+            roda.logger.info("Push new image '{}:{}' to {}:{} success".format(
+                new_image_name, new_image_tag, hub_addr, hub_port))
         except Exception as e:
             for line in docker_resp:
                 roda.logger.error(line)
